inference/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
ultralytics==8.4.19
Pillow==10.4.0
numpy==1.26.4
pydantic==2.9.2
python-multipart==0.0.12
requests==2.32.3
torch==2.5.1
torchvision==0.20.1
